db bug fix

+++ app/db.py
@@ -13,4 +13,4 @@ class Users(db.Model):
     userid = db.Column(db.String(50), nullable=False)
     theme = db.Column(db.String(50), nullable=False)
     auth = db.Column(db.String(10), nullable=False)
-    phone = db.Column(db.Integer(20), nullable=False)
+    phone = db.Column(db.Integer, nullable=False)
